+++ src/main/java/org/realityforge/bazel/depgen/metadata/RecordUtil.java
@@ -15,6 +15,7 @@ import java.util.ArrayList;
 import java.util.Base64;
 import java.util.Map;
 import java.util.jar.JarFile;
+import java.util.stream.Collectors;
 import java.util.zip.ZipEntry;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -128,6 +129,35 @@ final class RecordUtil
               }
             }
           }
+        }
+      }
+      catch ( final IOException ignored )
+      {
+        // Fall through
+      }
+    }
+
+    return DepgenMetadata.SENTINEL;
+  }
+
+  @Nonnull
+  static String readJsAssets( @Nonnull final File file )
+  {
+    if ( isJarFile( file ) )
+    {
+      try
+      {
+        try ( final JarFile jar = new JarFile( file ) )
+        {
+          final String assetList =
+            jar
+              .stream()
+              .filter( e -> !e.isDirectory() )
+              .map( ZipEntry::getName )
+              .filter( name -> name.endsWith( ".js" ) && !name.contains( "/public/" ) )
+              .sorted()
+              .collect( Collectors.joining( "," ) );
+          return assetList.isEmpty() ? DepgenMetadata.SENTINEL : assetList;
         }
       }
       catch ( final IOException ignored )