+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Sort `omit_*` parameters by symbol in macro declarations.
 * Avoid adding a workspace rule for an artifact with no classifier unless the artifact has the "Java" or "Plugin" nature. Artifacts that only have the "J2cl" nature use the `sources` classifier and thus the workspace rule sans classifier is unnecessary.
 * Update the bazel version tested against to `3.3.0`.
+* Artifacts with a "J2cl" nature that contain *.js files will have the files added to the `j2cl_library` target. The exception is files nested within a directory named "public" as it is assumed that these are js files that are shipped as assets in GWT applications.
 
 ### [v0.10](https://github.com/realityforge/bazel-depgen/tree/v0.10) (2019-10-11) · [Full Changelog](https://github.com/realityforge/bazel-depgen/compare/v0.09...v0.10)